Update opt object on model date set
When setStartDate/setEndDate is called (when the model date is set) the internal opts object should update so it should "remember" this change when re-initing.

Currently, without this change, when the options are set from the template it triggers the "opts" watcher, which calls init - and loses the currently selected start/end date.